Final pre-dissertation changes
+Added spatie activity log
+Added activity log for all relevant actions
+Added route for accessing logs
+Fixed ide_helper installation

+++ app/Http/Controllers/Auth/AuthController.php
@@ -8,6 +8,8 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 
+// use Spatie\Activitylog\Models\Activity;
+
 class AuthController extends Controller
 {
 
@@ -17,7 +19,8 @@ class AuthController extends Controller
     public function register(Request $request)
     {
         if (Auth::user()->username != "admin") {
-            return response(null, 401);
+            activity("auth_fail")->log("Intento de acceso no autorizado a cliente: ", $request->ip);
+            abort(401);
         }
         $validatedData = $request->validate([
             "username" => "required|string|max:50|unique:users",
@@ -34,6 +37,12 @@ class AuthController extends Controller
         ]);
         $token = $user->createToken("auth_token")->plainTextToken;
 
+        activity("register")->log(
+            "Analista " .
+            $validatedData["name"] .
+            " registrado como usuario " .
+            $validatedData["username"]);
+
         return response()->json(
             [
                 "access_token" => $token,
@@ -45,11 +54,14 @@ class AuthController extends Controller
     public function login(Request $request)
     {
         if (!Auth::attempt($request->only(["username", "password"]))) {
+            activity("auth_fail")->log("Intento de acceso no autorizado a cliente " . getHostByName(getHostName()));
+
             return response()->json(["message" => "invalid"], 401);
         }
         $user = User::where("username", $request["username"])->firstOrFail();
         $token = $user->createToken("auth_token")->plainTextToken;
 
+        activity("login")->log($user->username);
         return response()->json([
             'access_token' => $token,
             'token_type' => 'Bearer',
@@ -60,18 +72,22 @@ class AuthController extends Controller
     }
     public function getUser(Request $request)
     {
-        return response()->json($request->user()->username);
+        if (!$request->user()) {
+            abort(401);
+        }
+        return $request->user();
 
     }
     public function logOut(Request $request)
     {
-        Auth::logout();
-        $request->session()->invalidate();
+        activity("logout")->log($request->user()->username);
 
     }
 
     public function deleteUser(Request $request)
     {
+        if (!$request->user()) {abort(401);}
+        activity("user_deletion")->log("Usuario " . $request->username . " Eliminado");
         User::where("username", $request->query("username"))->delete();
     }
